requirements.txt
requests>=2.28.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv